firebaseのエラーの消去
