+++ program/public/firebase.js
@@ -4,25 +4,20 @@
 // import { getFirestore } from "firebase/firestore"
 //require("dotenv").config({ debug: true });
 
+import firebase from 'firebase/compat/app';
+import 'firebase/compat/auth';
 
-// Your web app's Firebase configuration
-// const firebaseConfig = {
-//   apiKey: process.env.API_KEY,
-//   authDomain: process.env.AUTH_DOMAIN,
-//   projectId: process.env.PROJECT_ID,
-//   storageBucket: process.env.STORAGE_BUCKET,
-//   messagingSenderId: process.env.MESSAGING_SENDER_ID,
-//   appId: process.env.APP_ID
-// };
+const firebaseConfig = {
+  apiKey: process.env.API_KEY,
+  authDomain: process.env.AUTH_DOMAIN,
+  projectId: process.env.PROJECT_ID,
+  storageBucket: process.env.STORAGE_BUCKET,
+  messagingSenderId: process.env.MESSAGING_SENDER_ID,
+  appId: process.env.APP_ID
+};
 
-// // Initialize Firebase
-// // const firebaseapp = initializeApp(firebaseConfig);
-// // const auth = getAuth(firebaseapp);
-// // const db = getFirestore(firebaseapp);
-// // provider.addScope("repo");
+const firebaseapp = initializeApp(firebaseConfig);
+const auth = getAuth(firebaseapp);
+//const db = getFirestore(firebaseapp);
 
-// // export {auth, db}
-
-// firebase.initializeApp(firebaseConfig);
-
-// export const auth = firebase.auth();
+export {auth}
