+++ program/public/mainpage/main.js
@@ -1,23 +1,8 @@
-// import { initializeApp } from "firebase/app";
-// import { getAuth } from "firebase/auth";
+// import { initializeApp } from "/firebase/app";
+// import { getAuth } from "/firebase/auth";
 // //import { getFirestore } from "firebase/firestore"
-// import {signInWithEmailAndPassword, createUserWithEmailAndPassword} from "firebase/auth";
+// import {signInWithEmailAndPassword, createUserWithEmailAndPassword} from "/firebase/auth";
 
-// const firebaseConfig = {
-//   apiKey: process.env.API_KEY,
-//   authDomain: process.env.AUTH_DOMAIN,
-//   projectId: process.env.PROJECT_ID,
-//   storageBucket: process.env.STORAGE_BUCKET,
-//   messagingSenderId: process.env.MESSAGING_SENDER_ID,
-//   appId: process.env.APP_ID
-// };
-
-// const firebaseapp = initializeApp(firebaseConfig);
-// const auth = getAuth(firebaseapp);
-// //const db = getFirestore(firebaseapp);
-// provider.addScope("repo");
-
-import {auth} from "../app.js";
 
 //signup
 export async function createAccount() {
